Introduce SteadyBase. (#28819)
